fully intgrated AI into application #12 - edited certain DTOs to accept original model classes (to avoid StackOveflow errors when assigning or printing) - updated OpenAI API business logic to have better reasoning - prompt engineered user message validation, response to user input, and summarizing last 10 messages - updated some CSS for chat pane
- TODO: ask about if needing more than 1 summary is necessary

+++ backend/src/main/java/com/example/openai/dto/OpenAIRequest.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public class OpenAIRequest {
     @NotNull
-    private Integer recipeId;
+    private Long recipeId;
 
     @NotBlank
     @JsonProperty("message")
@@ -16,7 +16,7 @@ public class OpenAIRequest {
     public OpenAIRequest() {}
 
     // Getters and setters
-    public Integer getRecipeId() {
+    public Long getRecipeId() {
         return recipeId;
     }
